+++ bot/errors.py
@@ -7,7 +7,8 @@ __all__ = (
     "BadArgument",
     "BadAttachment",
     "NoAttachment",
-    "EmptyQueue"
+    "EmptyQueue",
+    "ItemNotFound"
 )
 
 
@@ -52,3 +53,23 @@ class NoAttachment(GenericError):
 class RoleError(GenericError):
     def __init__(self, role, author):
         super().__init__(f"'{role}' not in '{author}'.roles")
+
+
+class MoneyError(GenericError):
+    pass
+
+
+class ProfileAlreadyExists(MoneyError):
+    def __init__(self, user, msg=None):
+        msg = msg or f"{user} already has a profile"
+        super().__init__(msg)
+
+
+class ProfileNotFound(MoneyError):
+    def __init__(self, user, msg=None):
+        msg = msg or f"{user}'s profile doesn't exist, you can make one via `kc!money_start`"
+        super().__init__(msg)
+
+
+class ItemNotFound(MoneyError):
+    pass